==> notebooks/5.Other_plots.ipynb <==
## Plotting model $R^2$

import pandas as pd
import matplotlib.pyplot as plt

scores = {
    'NDM':pd.read_excel('../data/model_comparison.xlsx', sheet_name='NDM').set_index('cohort'),
    'linear':pd.read_excel('../data/model_comparison.xlsx', sheet_name='Linear').set_index('cohort'),
    'constant':pd.read_excel('../data/model_comparison.xlsx', sheet_name='Constant').set_index('cohort')
}

scores['NDM']

r2_cols = [col for col in scores['NDM'].columns if 'r2' in col]
std_cols = [col for col in scores['NDM'].columns if 'std' in col]

colors = ['#234990','#905023','#878220','#7e2390']
m=0

r2_cols = ['cortical-r2', 'braak5-r2', 'ndm-1-8-r2', 'ndm-1-4-r2', 'ndm-4-8-r2']
std_cols = ['cortical-std', 'braak5-std', 'ndm-1-8-std', 'ndm-1-4-std', 'ndm-4-8-std']

plt.figure(figsize=(8,6))
for model in scores.keys():
    plt.ylim([0.1,0.9])
    ad_r2 = scores[model][r2_cols].loc['AD_both_m4']
    ad_std = scores[model][std_cols].loc['AD_both_m4']
    plt.scatter(r2_cols, ad_r2, c=colors[m], s=120)
    plt.errorbar(r2_cols, ad_r2, ad_std, fmt=colors[m], mec=colors[m], ecolor=colors[m], ms=14, alpha=0.75)
    m += 1
plt.legend(scores.keys())
plt.show()

from cortography.utils import atlas_utils

# Load utility dataframe for DK regions
DK = atlas_utils.load_atlas('DK')

braak_5_regions = DK[ (DK['Braak stage']==5.0) & (DK['Cortex']=='cortical')]['Other Name']

braak_5_regions

colors = ['#234990','#878200','#7e2390']
m=0

r2_cols = ['cortical-r2', 'braak5-r2', 'ndm-1-8-r2', 'ndm-1-4-r2', 'ndm-4-8-r2']
std_cols = ['cortical-std', 'braak5-std', 'ndm-1-8-std', 'ndm-1-4-std', 'ndm-4-8-std']

plt.figure(figsize=(6,4))
markers=['o','v']
for model in ['NDM', 'linear']:
    plt.ylim([0.1,0.9])

    ad_r2 = scores[model][r2_cols].loc['AD_both_m4']
    ad_std = scores[model][std_cols].loc['AD_both_m4']
    plt.scatter(r2_cols, ad_r2, s=150, marker=markers[m], facecolors=colors[m], edgecolors=colors[m], linewidths=2, alpha=0.85)
    plt.errorbar(r2_cols, ad_r2, ad_std, fmt=colors[m])
    
    mci_r2 = scores[model][r2_cols].loc['MCI_both_m4'] + np.random.normal(0,0.01,5)
    mci_std = scores[model][std_cols].loc['MCI_both_m4']
    plt.scatter(r2_cols, mci_r2, s=150, marker=markers[m], facecolors='none', edgecolors=colors[m], linewidths=2, alpha=0.85)
    plt.errorbar(r2_cols, mci_r2, mci_std, linestyle = '--', fmt=colors[m])
    m += 1
plt.savefig('../results/notebook_figs/SI_r2.png', dpi=150, transparent=True, bbox_inches = 'tight', pad_inches = 0.1)
plt.show()

# Plotting $\Delta$$\tau$
This notebook plots the equation for $\Delta$$\tau$ as predicted by previous notebooks.

from mpl_toolkits.mplot3d import axes3d
from matplotlib import cm

### Plotting covariates

color_list = (AD_MCI_tau_v3 - AD_MCI_tau_v1)[NDM_predicted_regions].mean(axis=1).values

plt.figure(figsize=(10,6))
plt.xticks( fontsize=18)
plt.yticks( fontsize=18)
plt.xlabel(r'A$\beta_{baseline}$', fontsize=24)
plt.ylabel('FTP', fontsize=24)
# plt.xticks([])
# plt.yticks([])
plt.scatter(fbp.values, total_bl_tau.values, 
            c=color_list, 
            cmap = 'RdPu', 
            s=60, 
            alpha=0.75)
plt.colorbar()
plt.savefig('../results/figs/manuscript_figures/SI/ftp_vs_fbp.png', transparent=True, dpi=150)
plt.show()

color_list = (AD_MCI_tau_v3 - AD_MCI_tau_v1)[NDM_predicted_regions].mean(axis=1).values

plt.figure(figsize=(10,6))
plt.xticks( fontsize=18)
plt.yticks( fontsize=18)
plt.xlabel('FTP', fontsize=24)
plt.ylabel(r'FTP / A$\beta_{baseline}$', fontsize=24)
plt.ylim([-3,9])
# plt.yticks([])
plt.scatter(total_bl_tau.values, total_bl_tau.values/fbp.values, 
            c=color_list, 
            cmap = 'RdPu', 
            s=60, 
            alpha=0.75)
plt.colorbar()
plt.savefig('../results/figs/manuscript_figures/SI/ftp_vs_ratio.png', transparent=True, dpi=150)
plt.show()

color_list = (AD_MCI_tau_v3 - AD_MCI_tau_v1)[NDM_predicted_regions].mean(axis=1).values

plt.figure(figsize=(10,6))
plt.xticks( fontsize=18)
plt.yticks( fontsize=18)
plt.xlabel(r'A$\beta_{baseline}$', fontsize=24)
plt.ylabel(r'FTP / A$\beta_{baseline}$', fontsize=24)
plt.ylim([-3,9])
# plt.yticks([])
plt.scatter(fbp.values, total_bl_tau.values/fbp.values, 
            c=color_list, 
            cmap = 'RdPu', 
            s=60, 
            alpha=0.75)
plt.colorbar()
plt.savefig('../results/figs/manuscript_figures/SI/fbp_vs_ratio.png', transparent=True, dpi=150)
plt.show()

xy_dict = make_regression_variables(subjects='AD', model=2, regions_to_average=list(cortical_regions))
X,Y = xy_dict['X'], xy_dict['Y']

model = LinearRegression(fit_intercept=True, normalize=True)
model.fit(X, Y)

model.coef_.ravel()

model.intercept_[0]

fbp

a,b = list(model.coef_.ravel())
c = model.intercept_[0]

model.coef_.ravel()

amyloid = np.linspace(0.8,2,100)
tau = np.linspace(50,100,100)

X = np.array([[A for i in range(len(tau))] for A in amyloid])
Y = np.array([[t for t in tau] for A in amyloid])
Z = np.array([[-100*(a*t + b*A + c*(t/A)) + d for t in tau] for A in amyloid])

%matplotlib inline

# xx = fbp
# yy = total_bl_tau
# zz = [-100*(a*yy[i] + *yy[i] + c*(yy[i]/xx[i])) + d for i in range(len(fbp))]

%matplotlib widget
# aka import ipympl

import matplotlib.pyplot as plt

plt.plot([0, 1, 2, 2])
plt.show()

fig = plt.figure(figsize=(20,10))
ax = fig.gca(projection='3d')
# scat = ax.scatter(xx,yy,zz, s=400)
surf = ax.plot_surface(X, Y, Z, cmap=cm.coolwarm_r,
                       linewidth=0, antialiased=False,alpha=0.2)

AD_MCI_v1_df

## Plotting coefficients

model_test = make_regression_variables(subjects='AD', model=2, regions_to_average=list(cortical_regions))
X, Y = model_test['X'], model_test['Y']
bootstrapped_data = calculate_bootstrap_samples(X=X, Y=Y)

par0 = [bootstrapped_data['models'][i].coef_[0][0] for i in range(5000)]
par1 = [bootstrapped_data['models'][i].coef_[0][1] for i in range(5000)]
intercept = [bootstrapped_data['models'][i].intercept_[0] for i in range(5000)]


for i in range(len(fbp)):
    plt.scatter(fbp.values[i], total_bl_tau.values[i])
plt.hlines(y=0, xmin=-1.5, xmax=2.0)

for i in range(len(fbp)):
    plt.scatter(fbp.values[i], total_bl_tau.values[i]/fbp.values[i])

for i in range(len(fbp)):
    plt.scatter(total_bl_tau.values[i], total_bl_tau.values[i]/fbp.values[i])

for i in range(len(fbp)):
    plt.scatter(fbp.values[i]*total_bl_tau.values[i], total_bl_tau.values[i]*total_bl_tau.values[i])

color_list = (AD_MCI_tau_v3 - AD_MCI_tau_v1)[cortical_regions].sum(axis=1).values

plt.scatter(fbp.values, total_bl_tau.values, c=color_list, cmap = 'viridis')


_ = plt.hist(par0)
np.array(par0).mean()

_ = plt.hist(par1)
np.array(par1).mean()

_ = plt.hist(intercept)
np.array(intercept).mean()

_ = plt.hist(fbp, alpha=0.5)
_ = plt.hist(total_bl_tau, alpha=0.5)

dd

_ = plt.hist(fbp)
